[fix] DiaryDate 클래스의 위치를 presentation에서 domain:diary로 이동 및 이로 인해 기존 android 모듈에서만 사용할 수 있는 코드들을 제거 및 대체
