+++ presentation/src/main/java/com/strayalphaca/presentation/screens/diary/util/DiaryDateUtil.kt
@@ -1,0 +1,14 @@
+package com.strayalphaca.presentation.screens.diary.util
+
+import android.os.Build
+import com.strayalphaca.travel_diary.diary.model.DiaryDate
+
+fun DiaryDate.Companion.getInstanceFromDateString(
+    dateString : String, dateFormat : String = "yyyy-MM-dd"
+) : DiaryDate {
+    return if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+        getInstanceFromDateStringByDateTimeFormatter(dateString, dateFormat)
+    } else {
+        getInstanceFromDateStringBySimpleDateFormat(dateString, dateFormat)
+    }
+}
